Move `MockLegacyResponder` to its own file.
